Fix flakiness in SpeakerDeck module that sometimes results in slides not staying in sync with the video.
SpeakerDeck#changeSlide relied on having received a message from
SpeakerDeck's iframe before it would attempt to change to the specified
slide. SpeakerDeck usually posts this message when the first slide loads in the
iframe, but in some unknown circumstances we're not seeing it be
posted. It does not appear to be necessary to wait for the message from
SpeakerDeck since we're able to directly retrieve a reference to the
destination of the messages, SpeakerDeck's iframes' DOMWindow.